Update README.md
Rectangle.java
- Rectangle class is very basic, and simply calls for the 
creation of a rectangle with values l and w, length and width. It has 
methods to return the length and width of a certain rectangle.

ComparePerimeter.java
- An implementation of Comparator on the Rectangle class. This 
Class takes in two Rectangle objects and calculates their respective 
areas. If the first is greater, 1 is returned. If the second is 
greater then a -1 is returned.

CompareArea.java
- Exactly like ComparePerimeter except the formula is for area 
rather than perimeter.

Problem1.java
- The main class, which creates the rectangles, adds them to an 
Array of type Rectangle. Then, the findMax method is called on this 
array with first the area comparator and then the perimeter 
comparator as parameters and the value is set to the maxArea and 
maxPerimeter Rectangles. The length and width of the largest rectangles
are printed. The findMax method utilizes something similar to binary 
search, where it recursively searches through the entire array of 
Rectangles, comparing them by the set properties in ComparePerimeter 
and CompareArea. If a 1 is returned then the first value is greater 
and its index value is set as the maxIndex. The method returns the 
Rectangle which maxIndex is referencing.