Update README.md
Problem3Tester
- Class that takes the input of the array.txt file and inputs 
each line into an ArrayList of Integers. Has a method called a() that
converts the ArrayList into an int array so that it can be passed into
the binary search class.

Problem3BinSearch
- Class that recursively searches for the presence of a number 
when given an array list and a number to look for. The first search  
takes in two paramaters, and directs to the second search, where the
method takes in four parameters including a high and low value. By
finding the midpoint using the high and low values, then comparing 
that midpoint value to the value to search for, the algorithm 
eliminates half of the list with each iteration, and has a big O value 
of logn.
- If the value being searched for is greater than the middle   
value, then it's only possibly in the upper half of the remaining
list. Then it recursively performs the search again with the mid+1
value set as the lowest value. Vice versa for the opposite case.
The method performs recursively  until the middle value is the number
being searched for, or until the value of low is set to a greater
value than the high value, meaning the whole list has been searched,
in which case it returns -1.

Problem3
- The main class which throws exceptions for files that aren't
found and incorrect inputs. A try catch surrounds the code. The code 
runs continuously with a while loop.
- First Problem3Tester is called and the numbers from the array
text file are put into an Arraylist. Written commands ask the user to
input a value to be searched for.
- A scanner is instantiated, and the value input is taken in as
a double (just in case the user wants to quit, in which case they can
type in .5, which breaks out of the code and sets the while loop
boolean to false.) Otherwise, the value input is cast as an int.  
- Next Problem3BinSearch is called and we test the array above   
"test" by inputing the array as an int array (using the a() method to   
convert to an int array) and the user's number choice as parameters.
The answer is printed if found, and the user it told it isn't if a -1
is returned.
- Finally there are three catches, one for FileNotFound,
InputMismatch and NoSuchElement.