Update README.md
Bulls and Cows is a popular public domain code-breaking deduction game; this java program stimulates the game.
Four numbers are picked by the computer. The user guesses 4 numbers, and if the number is correct AND in the correct spot, the number gets counted as a "bull". If the number is correct but incorrectly placed, the number gets counted as a "cow". The user must continue to guess based on the response of "x cows, y bulls" until they correctly guess the 4 digit number.